requirements.txt
# requirements.txt
paho-mqtt==2.1.0
pymodbus==3.5.0
PyYAML==6.0.1
pyserial
